+++ src/server/include/common/cipher.h
@@ -3,6 +3,6 @@
 
 #define PAR_FOR_AT_K 256
 
-int cipher(char *input_file, char *output_file, unsigned long seed);
+int cipher(char *input_file, char *output_file, unsigned int seed);
 
 #endif // CIPHER_H
